+++ app/utils/camera/__init__.py
@@ -1,2 +1,14 @@
 from .camera_utils import CameraInterface
 from .segmentation_utils import segment_pcd_from_2d
+from .device.fake_camera import FakeCamera, FakeRGBDFrame
+from .device.http_camera import HTTPCamera
+from .device.usb_camera_parser import USBVideoManager
+
+__all__ = [
+    'CameraInterface',
+    'segment_pcd_from_2d',
+    'FakeCamera',
+    'FakeRGBDFrame',
+    'HTTPCamera',
+    'USBVideoManager'
+]
