+++ app/utils/camera/device/usb_camera_parser.py
