+++ app/viewers/__init__.py
@@ -1,8 +1,5 @@
 from .image_viewer import ResizableImageLabel
-from .scene_viewer import CameraReader, MultiCamStreamer, FakeCamera, FakeRGBDFrame
+from .scene_viewer import MultiCamStreamer
 
 __all__ = ['ResizableImageLabel', 
-           'FakeCamera', 
-           'FakeRGBDFrame', 
-           'CameraReader',
            'MultiCamStreamer']
