+++ app/viewers/scene_viewer.py
@@ -1,185 +1,9 @@
-import numpy as np
 import cv2
-import requests
-from typing import Union
-from PySide6.QtCore import QThread, QMutex
 from app.utils.logger import setup_logger
+from app.threads.cam_reader import CameraReader
+from app.utils.camera import HTTPCamera, FakeCamera
+
 logger = setup_logger(__name__)
-
-class FakeCamera:
-    """Fake camera that generates synthetic RGBD frames for debugging."""
-    def __init__(self):
-        self.width = 640
-        self.height = 480
-        self.frame_idx = 0
-
-    def connect(self, index):
-        """Fake connect method, always returns True."""
-        return True
-
-    def release(self):
-        pass
-
-    def disconnect(self):
-        """Fake disconnect method."""
-        pass
-
-    def read(self, both: bool = False):
-        color_image = np.zeros((self.height, self.width, 3), dtype=np.uint8)
-        center_x = int((self.frame_idx * 5) % self.width)
-        center_y = self.height // 2
-        cv2.circle(color_image, (center_x, center_y), 50, (0, 255, 0), -1)
-
-        # Generate a dynamic depth image
-        x = np.linspace(0, 2 * np.pi, self.width)
-        y = np.linspace(0, 2 * np.pi, self.height)
-        xx, yy = np.meshgrid(x, y)
-
-        # Use a sine wave to create dynamic depth variations
-        base_depth = 1000  # Base depth value in mm
-        amplitude = 500  # Amplitude of the depth variation
-        depth_image = base_depth + amplitude * np.sin(xx + self.frame_idx * 0.1)
-
-        # Convert to uint16
-        depth_image = depth_image.astype(np.uint16)
-
-        # Randomly zero out some regions to simulate missing depth
-        num_missing_regions = np.random.randint(5, 15)  # Random number of missing regions
-        for _ in range(num_missing_regions):
-            # Randomly choose the size and position of the missing region
-            start_x = np.random.randint(0, self.width - 50)
-            start_y = np.random.randint(0, self.height - 50)
-            width = np.random.randint(20, 100)
-            height = np.random.randint(20, 100)
-
-            # Zero out the region
-            depth_image[start_y:start_y + height, start_x:start_x + width] = 0
-        self.frame_idx += 1
-        if not both:
-            return True, color_image
-        return True, FakeRGBDFrame(depth_image, color_image)
-
-    def capture_frame(self, enable_align_depth_to_color=True):
-        """Generate synthetic depth and color images with dynamic depth regions."""
-        return self.read(both=True)[1]
-
-
-class FakeRGBDFrame:
-    """Fake RGBD frame containing synthetic depth and color images."""
-    def __init__(self, depth_image, color_image):
-        self.depth = depth_image
-        self.color = color_image
-
-
-class MJPEGStreamReader(QThread):
-    """Thread for reading MJPEG stream from HTTP cameras using PyQt QThread"""
-    
-    def __init__(self, url):
-        super().__init__()
-        self.url = url
-        self.running = True
-        self.frame = None
-        self.frame_lock = QMutex()
-        
-    def run(self):
-        try:
-            stream = requests.get(self.url, stream=True)
-            bytes_buffer = b""
-
-            for chunk in stream.iter_content(chunk_size=1024):
-                if not self.running:
-                    break
-                bytes_buffer += chunk
-                a = bytes_buffer.find(b'\xff\xd8')  # Start of JPEG
-                b = bytes_buffer.find(b'\xff\xd9')  # End of JPEG
-                if a != -1 and b != -1 and b > a:
-                    jpg = bytes_buffer[a:b+2]
-                    bytes_buffer = bytes_buffer[b+2:]
-                    img_array = np.frombuffer(jpg, dtype=np.uint8)
-                    frame = cv2.imdecode(img_array, cv2.IMREAD_COLOR)
-                    if frame is not None:
-                        self.frame_lock.lock()
-                        self.frame = frame
-                        self.frame_lock.unlock()
-        except Exception as e:
-            logger.error(f"Error in HTTP camera stream: {str(e)}")
-            self.running = False
-
-    def read(self):
-        self.frame_lock.lock()
-        if self.frame is None:
-            result = (False, None)
-        else:
-            result = (True, self.frame.copy())
-        self.frame_lock.unlock()
-        return result
-
-    def stop(self):
-        self.running = False
-        self.wait()  # Wait for the thread to finish
-
-
-class HTTPCamera:
-    """Represents an HTTP camera using MJPEG streaming"""
-    
-    def __init__(self, url):
-        self.url = url
-        self.stream_reader = MJPEGStreamReader(url)
-        self.stream_reader.start()
-        
-    def read(self):
-        return self.stream_reader.read()
-    
-    def release(self):
-        self.stream_reader.stop()
-        
-    def isOpened(self):
-        return self.stream_reader.isRunning()
-
-
-class CameraReader(QThread):
-    """Thread for reading frames from cameras continuously"""
-    
-    def __init__(self, camera_name, camera_capture: Union[cv2.VideoCapture, HTTPCamera]):
-        super().__init__()
-        self.camera_name = camera_name
-        self.camera = camera_capture
-        self.running = True
-        self.frame = None
-        self.frame_lock = QMutex()
-        self.frame_ready = False
-    
-    def run(self):
-        while self.running:
-            ret, frame = self.camera.read()
-            if ret:
-                frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-                
-                self.frame_lock.lock()
-                self.frame = frame
-                self.frame_ready = True
-                self.frame_lock.unlock()
-            else:
-                self.frame_lock.lock()
-                self.frame_ready = False
-                self.frame_lock.unlock()
-                
-            # Small delay to prevent CPU overuse
-            QThread.msleep(10)
-    
-    def get_frame(self):
-        self.frame_lock.lock()
-        if not self.frame_ready:
-            self.frame_lock.unlock()
-            return False, None
-        
-        frame_copy = self.frame.copy() if self.frame is not None else None
-        self.frame_lock.unlock()
-        return True, frame_copy
-    
-    def stop(self):
-        self.running = False
-        self.wait()  # Wait for the thread to finish
 
 
 class MultiCamStreamer:
@@ -190,6 +14,7 @@ class MultiCamStreamer:
     
     def get_frame(self, **kwargs):
         frames = {}
+        thread: CameraReader
         for camera_name, thread in self.camera_threads.items():
             ret, frame = thread.get_frame()
             if ret:
